커뮤니티 이전 메뉴 버튼 추가

+++ src/main/java/kr/kh/app/service/MemberService.java
@@ -16,10 +16,10 @@ import org.apache.ibatis.session.SqlSessionFactoryBuilder;
 import kr.kh.app.dao.MemberDAO;
 import kr.kh.app.model.vo.MemberVO;
 
-public class MemberService {
-	
+public class MemberService{
+
 	private MemberDAO memberDao;
-	
+
 	public MemberService() {
 		String resource = "kr/kh/app/config/mybatis-config.xml";
 		InputStream inputStream;
@@ -35,27 +35,36 @@ public class MemberService {
 		}
 	}
 
-	public boolean signUp(MemberVO member) {
-		if(member == null ||
-			member.getMe_id() == null ||
-			member.getMe_pw() == null ||
-			member.getMe_email() == null) return false;
-		
-		if (checkRegex(member.getMe_id(), "^\\w{6,13}$")) return false;
-		if (checkRegex(member.getMe_pw(), "^[a-zA-Z0-9!@#$%^&*()]{6,15}$")) return false;
-		if (checkRegex(member.getMe_email(), "^[A-Za-z0-9_]+@[A-Za-z0-9_]+(\\.[A-Za-z]{2,}){1,}$")) return false;
-		
+	public boolean signUp(String id, String pw, String pw_ckh, String email) {
+		if (checkRegex(id, "^\\w{6,13}$")) return false;
+
+		if (checkRegex(pw, "^(?=.*[A-Z])(?=.*[a-z])(?=.*[\\d])(?=.*[^\\w])([^\\w]{1}|[\\w]{1}){6,15}$")) {
+			return false;
+		}
+		if (checkRegex(pw_ckh, "^(?=.*[A-Z])(?=.*[a-z])(?=.*[\\d])(?=.*[^\\w])([^\\w]{1}|[\\w]{1}){6,15}$")) {
+			return false;
+		}
+		if (checkRegex(email, "^[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*\\.[a-zA-Z]{2,3}$")) {
+			return false;
+		}
+		if (!pw.equals(pw_ckh)) {
+			return false;
+		}
+
+		MemberVO newUser = new MemberVO(id, pw, email);
 		try {
-			return memberDao.insertMember(member);
+			memberDao.insertMember(newUser);
 		} catch (Exception e) {
 			e.printStackTrace();
 			return false;
 		}
+		return true;
 	}
 
 	private boolean checkRegex(String str, String regex) {
-		if (str == null || str.trim().length() == 0) return true;
+		if (str == null || str.trim().length() == 0) return false;
 		if (Pattern.matches(regex, str)) return false;
+
 		return true;
 	}
 
@@ -68,21 +77,17 @@ public class MemberService {
 		if (member == null) return null;
 
 		MemberVO user = memberDao.selectMember(member.getMe_id());
-
 		// 가입되지 않은 아이디이면
-		if (user == null) {
-			return null;
-		}
+		if (user == null) return null;
 		// 비번이 같으면
-		if (user.getMe_pw().equals(member.getMe_pw())) {
-			return user;
-		}
+		if (user.getMe_pw().equals(member.getMe_pw())) return user;
 		return null;
 	}
 	
 	public Cookie createCookie(MemberVO user, HttpServletRequest request) {
-		if (user == null) return null;
-		
+		if (user == null) {
+			return null;
+		}
 		HttpSession session = request.getSession();
 		// 쿠키는 이름, 값, 만료시간, path가 필요
 		String me_cookie = session.getId();
@@ -104,19 +109,22 @@ public class MemberService {
 	}
 
 	public void updateMemberCookie(MemberVO user) {
+		if(user == null) return;
 		memberDao.updateMemberCookie(user);
 	}
 
+	public void deletmemeber(MemberVO user) {
+		if(user == null) return;
+		memberDao.deleteMember(user);
+	}
+
 	public boolean updateMemberPw(MemberVO user, String newpw) {
+		if(user == null || newpw == null || newpw.trim().length() == 0) return false;
 		return memberDao.updateMemberPw(user, newpw);
 	}
 
-	public void deletmemeber(MemberVO user) {
-		memberDao.deleteMember(user);
-		
-	}
-
 	public boolean updateMemberEmail(MemberVO user, String newemail) {
+		if(user == null || newemail == null || newemail.trim().length() == 0) return false;
 		return memberDao.updateMemberEmail(user, newemail);
 	}
